Close #8: Internationalize player
Actually, the video.js player is already internationalized and several
language files are already available. Instead of using these language
files directly, we integrate the localization in the CMSimple_XH plugin
language files. This makes it much easier to add missing translations
(or to change the existing translations), and also nicely works around
some mapping issues regarding the language codes (for instance, zh-TW).

For the hopefully unlikely case that json_encode() is not available, we
fall back to the default language strings.