Remove HtmlString
The whole concept is flawed.  The first problem is `::__toString()`
which easily defeats the purpose of type safety.  Then we realize that
`View::escape()` would need to return an `HtmlString`, but without the
magic `::__toString()` we would need yet another method to actually get
the string, so `$this->esc($variable)->toString()`?  That doesn't look
like being worth the trouble.

Instead, we say goodbye to semi-automatic escaping, and introduce
`View::raw()` for cases where it is needed.  We also rename `::escape()`
for symmetry and brevity.