Don't upscale both video dimensions
That would only make sense if we had a sophisticated upscaling
algorithm, but as is, this is just a waste of space.

We still allow upscaling a single dimension, though, so that we can
have proper 480p versions of PAL/NTSC 16:9 videos.

The smallest input size we're going to support is 240p.  Even that is
very conservative – almost nobody wants to watch such constrained
videos.