Use VP9 for WebM
Without the INI files, WebM videos are preferred over MP4, what makes
sense.  However, VP8 isn't as good as MP4, so we switch to VP9; while
the encoding is much slower (what isn't a big deal if we encode only
a single resolution, anyway), we get more bang for the buck.