Drop useless copyright years
Usually, copyright protection is not coupled to the creation time of a
software, but rather to the lifetime of the creator.  As such, the
copyright years are pretty useless, and since they are hard to maintain
(we cannot just update these at the start of new year), we just drop
them.

For the tests, we drop the license headers altogether.