Fix #43: Vertical videos have bad quality
In case a vertical (aka. portrait mode) video is passed to ITube, we
resize to optimize for 9:16 screens, and the maximum resolution is 480x
854 (what we call 480vp).  The reason for not encoding 720vp is that on
a FullHD screen (1920x1080) a 720p video will be stretched in either
dimension by a factor of 1.5, while a 720vp video will be stretched by
0.844 (i.e. downscaled).  So a 720vp video would be overkill for a
typical desktop monitor (or TV screen).  For mobiles which nowadays may
have even higher resolution than typical desktop monitors, 480vp still
appears to be sufficient, since these devices have way more DPI (likely
x3 or x4), so the higher video resolution might not even be noticed at
all.  And while nowadays the Web is more often used on mobiles, desktop
is still pretty common.  The only drawback with this algorithm would be
tablets and vertically set up desktop monitors, but either is rather
uncommon.  And in the end we need to make compromises.