Lay out call builder as table
That was actually what we had in mind, but the inline-blocks with
widths don't necessarily work.  Thus we switch to a table layout.

A grid layout would provide more flexibility for users, but we're
reluctant to only rely on CSS3 features, and it's somewhat unlikely
that users would want to fiddle with the grid layout in this case,
anyway.